+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Expose an interface path from Astarte aggregate datastream event.
+
 ## [1.1.0] - 2023-01-16
 
 ## [1.1.0-alpha.0] - 2022-12-15